Minigame Score
made the score calculations of the minigames to be readable. this will average the score the player gained in the minigames

+++ Assets/Scripts/Minigames/IceMinigame/IceTowerGameManager.cs
@@ -155,7 +155,7 @@ public class IceTowerGameManager : MonoBehaviour
     {
         if (isGameOver)
             return;
-        iceQualityScore = currentScore / targetScore;
+        iceQualityScore = (float)currentScore / (float)targetScore;
         IceGameEnded.Raise(this.gameObject, MinigameType.Ice);
         isGameOver = true;
         print("GameOver");
